+++ packages/open-next/CHANGELOG.md
@@ -1,5 +1,47 @@
 # open-next
 
+## 0.9.3
+
+### Patch Changes
+
+- a574834: Support API Gateway REST API event
+- 81dbb69: Support Next.js apps without "public" folder
+
+## 0.9.2
+
+### Patch Changes
+
+- ea0ab0c: Fix requesting public files returns 404
+- 0f56b91: Add strict checking to detect public file request
+
+## 0.9.1
+
+### Patch Changes
+
+- 0dadccb: Set default NODE_ENV to production
+
+## 0.9.0
+
+### Minor Changes
+
+- 787c1b2: Support Next.js 404 pages
+
+### Patch Changes
+
+- 6395849: Fix "Cannot find package 'next'"
+
+## 0.8.2
+
+### Patch Changes
+
+- e5b5204: Handle `next build` error
+
+## 0.8.1
+
+### Patch Changes
+
+- bdd29d1: Fix spawn error on Windows
+
 ## 0.8.0
 
 ### Minor Changes